Added Create And Delete Validation
Can't Delete a tenant That Does Not Exists
Can't add a tenant that already exists

+++ src/main/java/com/qa/tenantApi/main/service/TenantService.java
@@ -40,11 +40,18 @@ public class TenantService {
 	}
 
 	public String createTenant(Tenant tenant) {
+		if(this.tenantRepo.getTenantByFirstNameAndLastName(tenant.getFirstName(), tenant.getLastName()) != null) {
+			return "Tenant Already Exists";
+		}
 		this.tenantRepo.save(tenant);
 		return Constants.getCreationMessage();
 	}
 
 	public String deleteTenant(Tenant tenant) {
+		if(this.tenantRepo.getTenantByFirstNameAndLastName(tenant.getFirstName(), tenant.getLastName()) == null)
+		{
+			return "Tenant Does Not Exist";
+		}
 		this.tenantRepo.delete(tenant);
 		return Constants.getDeletionMessage();
 	}
